// src/services/api/models/profile-response-data.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Documentation
 * api documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ProfileResponseData
 */
export interface ProfileResponseData {
    /**
     * 
     * @type {number}
     * @memberof ProfileResponseData
     */
    'id': number;
    /**
     * 
     * @type {string}
     * @memberof ProfileResponseData
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof ProfileResponseData
     */
    'firstName': string;
    /**
     * 
     * @type {string}
     * @memberof ProfileResponseData
     */
    'middleName': string;
    /**
     * 
     * @type {string}
     * @memberof ProfileResponseData
     */
    'lastName': string;
    /**
     * 
     * @type {string}
     * @memberof ProfileResponseData
     */
    'birthDate': string;
    /**
     * 
     * @type {string}
     * @memberof ProfileResponseData
     */
    'height': string;
    /**
     * 
     * @type {string}
     * @memberof ProfileResponseData
     */
    'weight': string;
    /**
     * 
     * @type {string}
     * @memberof ProfileResponseData
     */
    'gender': string;
    /**
     * 
     * @type {string}
     * @memberof ProfileResponseData
     */
    'role': string;
    /**
     * 
     * @type {string}
     * @memberof ProfileResponseData
     */
    'updatedAt': string;
    /**
     * 
     * @type {string}
     * @memberof ProfileResponseData
     */
    'createdAt': string;
}